feat: add more members' info
